backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
PyJWT==2.10.1
bcrypt==4.2.1
email-validator==2.2.0
pytest==8.3.4
alembic==1.14.1
